+++ frame-boost/src/main.rs
@@ -64,8 +64,8 @@ struct Args {
     n_frames: Option<u32>,
 
     /// Number of seconds to encode for scene. Higher value increase the confidence than all the frames in the scene will be above your quality target at cost of encoding time
-    #[arg(short = 's', long = "n-frames", default_value_t = 1, value_parser = clap::value_parser!(u32).range(1..))]
-    s_frames: u32,
+    #[arg(short = 's', long = "n-frames", default_value_t = 1.0)]
+    s_frames: f64,
 
     /// Workers to use when encoding
     #[arg(short = 'w', long, default_value_t = 2, value_parser = clap::value_parser!(u32).range(1..))]
